+++ logprep/abc/processor.py
@@ -150,7 +150,7 @@ class Processor(Component):
         self._result = value
 
     @property
-    def rules(self):
+    def rules(self) -> list["Rule"]:
         """Returns all rules
 
         Returns